+++ src/main.js
@@ -1,14 +1,14 @@
 import './assets/main.css'
 
 import { createApp } from 'vue'
-import { createPinia } from 'pinia'
+// import { createPinia } from 'pinia'
 
 import App from './App.vue'
-// import router from './router'
+import router from './router'
 
 const app = createApp(App)
 
-app.use(createPinia())
-// app.use(router)
+// app.use(createPinia())
+app.use(router)
 
 app.mount('#app')
